--- packages/expo-dev-launcher/android/src/main/java/expo/modules/devlauncher/helpers/DevLauncherOkHttpExtension.kt
@@ -1,36 +1,0 @@
-package expo.modules.devlauncher.helpers
-
-import android.net.Uri
-import kotlinx.coroutines.suspendCancellableCoroutine
-import okhttp3.Call
-import okhttp3.Callback
-import okhttp3.Headers
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.Response
-import java.io.IOException
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-
-/**
- * An extension which converts the OkHttp requests to coroutines.
- */
-suspend inline fun Request.await(okHttpClient: OkHttpClient): Response {
-  return suspendCancellableCoroutine { callback ->
-    okHttpClient.newCall(this).enqueue(object : Callback {
-      override fun onResponse(call: Call, response: Response) {
-        callback.resume(response)
-      }
-
-      override fun onFailure(call: Call, e: IOException) {
-        if (callback.isCancelled) {
-          return
-        }
-        callback.resumeWithException(e)
-      }
-    })
-  }
-}
-
-fun fetch(url: Uri, method: String, headers: Headers) =
-  Request.Builder().method(method, null).url(url.toString()).headers(headers).build()
