+++ packages/expo-dev-launcher/android/src/main/java/expo/modules/devlauncher/logs/DevLauncherRemoteLog.kt
@@ -8,22 +8,23 @@ internal interface DevLauncherRemoteLogBody {
   val message: String
   val stack: String?
 
-  fun toJson(): String
+  override fun toString(): String
 }
 
 internal class DevLauncherSimpleRemoteLogBody(override val message: String) : DevLauncherRemoteLogBody {
   override val stack: String? = null
 
-  override fun toJson(): String = message
+  override fun toString(): String = message
 }
 
 internal class DevLauncherExceptionRemoteLogBody(exception: Throwable) : DevLauncherRemoteLogBody {
   override val message: String = exception.toString()
   override val stack: String = exception.stackTraceToRemoteLogString()
 
-  override fun toJson(): String = Gson().toJson(this)
+  override fun toString(): String = Gson().toJson(this)
 }
 
+@Suppress("UNUSED")
 internal data class DevLauncherRemoteLog(
   val logBody: DevLauncherRemoteLogBody,
   @Expose val level: String = "error"
@@ -31,9 +32,8 @@ internal data class DevLauncherRemoteLog(
   @Expose
   val includesStack = logBody.stack !== null
 
-  @Suppress("UNUSED")
   @Expose
-  private val body = logBody.toJson()
+  private val body = logBody.toString()
 
   fun toJson(): String {
     return GsonBuilder()
